added isPermitted function to call _permitted
